pyproject.toml
[project]
name = "dicom-nlquery"
version = "0.1.0"
description = "Natural language DICOM query tool"
requires-python = ">=3.12"
dependencies = [
  "pynetdicom>=2.1.1",
  "pydicom>=2.4.0",
  "pydantic>=2.0",
  "httpx>=0.28.0",
  "pyyaml>=6.0.0",
  "click>=8.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4.0",
  "pytest-cov>=4.0.0",
]

[project.scripts]
dicom-nlquery = "dicom_nlquery.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
